Updated test structure
