+++ src/SALib/sample/radial/__init__.py
@@ -1,0 +1,2 @@
+# By default - forward `radial_sobol` interface
+from .radial_sobol import cli_parse, cli_action
